// alerts/apps/frontend/src/components/datatable/datatable.type.ts
export interface DataTableColumn<T> {

	/**
	 * Column accessor.
	 * You can use dot-notation for nested objects property drilling.
	 * (i.e. `department.name` or `department.company.name`)
	 */
	accessor: keyof T | (string & {})

	/**
	 * Center the column content.
	 * @default false
	 */
	center?: boolean

	/**
	 * Whether the column is hidden.
	 * @default false
	 */
	hidden?: boolean

	/**
	 * Custom cell data render function.
	 * Accepts the current record and its index in `records` as arguments and returns a React node
	 * (remember that a string is a valid React node too).
	 *
	 * Will override the default render function for this column.
	 */
	render?: (record: T) => React.ReactNode

	/**
	 * Whether the column is sortable.
	 * @default false
	 */
	sortable?: boolean

	/**
	 * The key to be used for sorting.
	 * If not provided, the column accessor will be used.
	 */
	sortKey?: string

	/**
	 * Optional column header title.
	 * If not present, one will be generated by "humanizing" the provided column accessor.
	 * (i.e. `firstName` -> `First name`; `user.firstName` -> `User first name`)
	 */
	title?: React.ReactNode | string

	/**
	 * The width of the column in pixels.
	 * @default 100%
	 */
	width?: number

}